Fixed the SGML <htmlurl> and <url> tags.
Those tags have two attributes: "url=" and "name=".  In the non-HTML output formats, <url> shows both fields, while <htmlurl> shows only the name field (as the HTML format always does.)  Thus, the general rules are
1. If the two attributes are different, then use <url>.
2. If they are similar, then use <htmlurl>.
3. If they are the same, then consider using <url> without the "name=" attribute.
(The reason for rules 2 and 3 is that the same text shouldn't be shown twice.)

There can be exceptions.  Example: "od65.sgml" has <htmlurl> because the URL
would disturb the flow of a sentence.