Guard against ld65 crashes when LineInfos is empty
Certain code paths in error handling can attempt to get source line info with empty LineInfos lists. This causes a precondition failure on the collection lookup.

The specific code path that was triggering this for me was a data segment containing a single condes table as the only data, meaning that there were no lines in the LineInfos collection. When the error handling in segments.c for SEG_EXPR_RANGE_ERROR (and others) triggered on the segment (for unrelated reasons) I got a precondition failure instead of a helpful error message. Other cases that create segments with no lines would also likely trigger similar errors elsewhere.