docs: Move intermediate files for building PDFs into 'docwrk' directory.
Another change is that the Makefile is now more intelligent about how
many 'latex' runs are needed. This is done by comparing the latex log
file of the previous run with the log file of the current run. This
change needs additional tools on the host: 'expr' and 'md5sum'

Some other cleanups/improvements to doc/Makefile.

Update .gitignore with 'pdf' and 'docwrk' directory entries.