vic20/cputc: Fix incorrect CRAM_PTR at startup when using conio
To do this we add a constructor call to the routine that fixes up
CRAM_PTR to match the screen location pointed to by SCREEN_PTR. We
also add a symbol for this, UPDCRAMPTR, and use that in vic20/kplot.s
where it is also called. (We also clarify a comment there.)

This adds two additional bytes to programs using cputc() or other
routines that call it.

Thanks to <greg.king5@verizon.net> (GitHub: greg-king5) for
investigating the difference in the VIC-20 KERNAL from the C64 and
proposing this solution to the problem.[1]

[1]: https://github.com/cc65/cc65/issues/946#issuecomment-538502820